extract component and move logic
